pyproject.toml
[project]
name = "website-developer-bot"
version = "0.1.0"
description = "WhatsApp bot that updates a website via Claude API, deployed on Cloudflare Workers"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.0",
    "requests>=2.31.0",
    "flask>=3.0.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
dev = "local_server:app.run"
